Added more animations and improved mobile compatibility.
