+++ src/styles/ArticleMainSectionLoneliness.module.css
@@ -16,44 +16,53 @@ a {
   text-decoration: underline; 
 }
 
-.sectionTitle > h3{
-	text-align: center;
-	position: relative;
-	font-family: 'Cherry Swash';
-  	font-size: 3.5rem; /* Implemented a slightly smaller font-size */
-  	color: #E07329;
-	line-height: 100%;
-	z-index: 1;
-	margin-top: 5rem;
+.sectionTitle > img{
+	margin: 5rem auto 0 auto;
+	display: block;
+	max-width: 100%;
+	align-items: center;
 }
 
-.sectionTitle > h3::before {
-	content: attr(data-text);
-	text-align: center;
-	z-index: -1;
-	position: absolute;
-	top: 0px;
-	left: 0px;
-	-webkit-text-stroke-color: #F9F4EB;
-	-webkit-text-stroke-width: .75rem;
-	text-shadow:
-		0 9px 12px rgb(0, 0, 0, 0.25),
-		-5px 9px 12px rgb(0, 0, 0, 0.25),
-		5px 9px 12px rgb(0, 0, 0, 0.25)
-}
 
 .heartCatWrapper {
 	display: flex;
+	position: relative;
 	justify-content: center;
 	align-items: center;
-	margin: 3rem auto;
+	margin: 3.5rem auto 2rem auto;
+	height: 20rem;
 	width: 100%;
 }
 
-.heartCatWrapper img {
+.heartCatWrapper > img {
+	position: absolute;
 	display: block;
-	max-width: 100%;
-	height: auto;
+	opacity: 0;
+}
+
+.heartCatWrapper > img:nth-of-type(1) {
+	filter: opacity(0.3);
+	bottom: 25px;
+}
+
+.heartCatWrapper > img:nth-of-type(2) {
+	filter: opacity(0.5);
+	bottom: 20px;
+}
+
+/* Images of first Heart Wrapper made bigger */
+.firstHeartCatWrapper > img:nth-of-type(1) {
+	scale: 1.08;
+}
+.firstHeartCatWrapper > img:nth-of-type(2) {
+	scale: 0.92;
+}
+
+.secondHeartCatWrapper > img:nth-of-type(1) {
+	scale: 1;
+}
+.secondHeartCatWrapper> img:nth-of-type(2) {
+	scale: 0.85;
 }
 
 .buildingContainer{
@@ -61,8 +70,8 @@ a {
 	display: block;
 	padding: 0px;
 	width: 100%;
-	height: 280px;
-	margin: 2rem auto 12rem auto;
+	height: 15rem;
+	margin: 0rem auto 12rem auto;
 }
 
 .buildingContainer > img {
@@ -70,8 +79,6 @@ a {
 	bottom: 0px;
 	margin-bottom: 0;
 	margin-top: 0;
-
-	
 }
 
 /* BACKGROUND GRASS */
@@ -84,20 +91,20 @@ a {
 /* SCHMITT HALL BUILDING */
 .buildingContainer > img:nth-of-type(2){
 	left: 50%;
-	bottom: -42px;
-	transform: translateX(-50%);
+	bottom: -4rem;
+	transform: translateX(-50%) scale(85%);
 	filter: drop-shadow(0 4px 4px rgba(0, 0, 0, 0.25));
 }
 
 /* LEFT BUSH */
 .buildingContainer > img:nth-of-type(3){
-	left: 0;
+	left: -15px;
 	transform: translateX(-50%);
 }
 
 /* RIGHT BUSH */
 .buildingContainer > img:nth-of-type(4){
-	right: 0;
+	right: -15px;
 	transform: translateX(50%);
 }
 
@@ -117,6 +124,7 @@ a {
 .smallQuoteContainer{
 	margin: 3rem 0rem;
 }
+
 .quote {
 	padding: 1.5rem;
 	border-radius: 2.5rem;
@@ -124,10 +132,14 @@ a {
 	border: 1rem solid #d3dbbc;
 }
 
+.smallQuoteContainer > p {
+	margin-top: 1rem;
+}
+
 .highFiveContainer{
 	position: relative; 
 	margin: auto;
-	width: 17rem;
+	max-width: 17rem;
 	height: 20rem;
 	
 }
@@ -146,29 +158,27 @@ a {
 
 .highFiveContainer.active > img:nth-of-type(1) {
 	animation: highFiveHeart 2s 2;
-	animation-delay: 2s;
+	animation-delay: 1.5s;
 }
 
 /* HAND */
 .highFiveContainer > img:nth-of-type(2) {
-	right: -1rem;
+	left: 50%;
 	transform-origin: bottom right;
 }
 
 .highFiveContainer.active > img:nth-of-type(2){
 	animation: highFiveHand 3s;
-	animation-delay: 1s;
 }
 
 /* PAW */
 .highFiveContainer > img:nth-of-type(3) {
-	left: 0;
+	right: 50%;
 	transform-origin: bottom left;
 }
 
 .highFiveContainer.active > img:nth-of-type(3) {
 	animation: highFivePaw 3s;
-	animation-delay: 1s;
 }
 
 .highFiveWrapper > p {
@@ -180,7 +190,7 @@ a {
 }
 
 .highFiveWrapper.active > p {
-	animation: highFiveText .75s ease-out 1.5s forwards;
+	animation: highFiveText .75s ease-out 1s forwards;
 }
 
 @keyframes highFiveHeart {
@@ -215,7 +225,7 @@ a {
 		transform: translate(-5px, 0px) rotate(0deg); /* return easing */
 	}
 	100% {
-		transform: translate(0px, 0px) rotate(0deg); /* fully back to start ✅ */
+		transform: translate(0px, 0px) rotate(0deg); /* fully back to start */
 	}
 }
 
@@ -236,7 +246,7 @@ a {
 		transform: translate(5px, 0px) rotate(0deg); /* return easing */
 	}
 	100% {
-		transform: translate(0px, 0px) rotate(0deg); /* fully back to start ✅ */
+		transform: translate(0px, 0px) rotate(0deg); /* fully back to start */
 	}
 }
 
@@ -251,7 +261,6 @@ a {
 		transform: translateY(0px);
 	}
 }
-
 
 .darkBrownEllipses{
 	position: absolute;
